Extract web authentication logic from domain layer to infrastructure layer
