+++ src/main/java/uib/swarchitecture/quepasa/infrastructure/web/controllers/UserController.java
@@ -1,4 +1,4 @@
-package uib.swarchitecture.quepasa.infrastructure.controllers;
+package uib.swarchitecture.quepasa.infrastructure.web.controllers;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
